requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.0
importlib-resources==5.4.0
pkg-resources==0.0.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.1.1
zipp==3.6.0
